+++ lib/Channel.d.ts
@@ -5,7 +5,7 @@ export declare class Channel extends EnhancedEventEmitter {
     private readonly _consumerSocket;
     private _nextId;
     private readonly _sents;
-    private _recvBuffer?;
+    private _recvBuffer;
     /**
      * @private
      */