+++ src/Channel.ts
@@ -1,10 +1,10 @@
+import * as os from 'os';
 import { Duplex } from 'stream';
-// @ts-ignore
-import * as netstring from 'netstring';
 import { Logger } from './Logger';
 import { EnhancedEventEmitter } from './EnhancedEventEmitter';
 import { InvalidStateError } from './errors';
 
+const littleEndian = os.endianness() == 'LE';
 const logger = new Logger('Channel');
 
 type Sent =
@@ -17,9 +17,9 @@ type Sent =
 	close: () => void;
 }
 
-// netstring length for a 4194304 bytes payload.
-const NS_MESSAGE_MAX_LEN = 4194313;
-const NS_PAYLOAD_MAX_LEN = 4194304;
+// Binary length for a 4194304 bytes payload.
+const MESSAGE_MAX_LEN = 4194308;
+const PAYLOAD_MAX_LEN = 4194304;
 
 export class Channel extends EnhancedEventEmitter
 {
@@ -39,7 +39,7 @@ export class Channel extends EnhancedEventEmitter
 	private readonly _sents: Map<number, Sent> = new Map();
 
 	// Buffer for reading messages from the worker.
-	private _recvBuffer?: Buffer;
+	private _recvBuffer = Buffer.alloc(0);
 
 	/**
 	 * @private
@@ -66,7 +66,7 @@ export class Channel extends EnhancedEventEmitter
 		// Read Channel responses/notifications from the worker.
 		this._consumerSocket.on('data', (buffer: Buffer) =>
 		{
-			if (!this._recvBuffer)
+			if (!this._recvBuffer.length)
 			{
 				this._recvBuffer = buffer;
 			}
@@ -77,76 +77,79 @@ export class Channel extends EnhancedEventEmitter
 					this._recvBuffer.length + buffer.length);
 			}
 
-			if (this._recvBuffer!.length > NS_PAYLOAD_MAX_LEN)
+			if (this._recvBuffer.length > PAYLOAD_MAX_LEN)
 			{
-				logger.error('receiving buffer is full, discarding all data into it');
+				logger.error('receiving buffer is full, discarding all data in it');
 
 				// Reset the buffer and exit.
-				this._recvBuffer = undefined;
+				this._recvBuffer = Buffer.alloc(0);
 
 				return;
 			}
 
+			let msgStart = 0;
+
 			while (true) // eslint-disable-line no-constant-condition
 			{
-				let nsPayload;
+				const readLen = this._recvBuffer.length - msgStart;
 
-				try
+				if (readLen < 4)
 				{
-					nsPayload = netstring.nsPayload(this._recvBuffer);
-				}
-				catch (error)
-				{
-					logger.error(
-						'invalid netstring data received from the worker process: %s',
-						String(error));
-
-					// Reset the buffer and exit.
-					this._recvBuffer = undefined;
-
-					return;
+					// Incomplete data.
+					break;
 				}
 
-				// Incomplete netstring message.
-				if (nsPayload === -1)
-					return;
+				const dataView = new DataView(
+					this._recvBuffer.buffer,
+					this._recvBuffer.byteOffset + msgStart);
+				const msgLen = dataView.getUint32(0, littleEndian);
+
+				if (readLen < 4 + msgLen)
+				{
+					// Incomplete data.
+					break;
+				}
+
+				const payload = this._recvBuffer.subarray(msgStart + 4, msgStart + 4 + msgLen);
+
+				msgStart += 4 + msgLen;
 
 				try
 				{
 					// We can receive JSON messages (Channel messages) or log strings.
-					switch (nsPayload[0])
+					switch (payload[0])
 					{
-						// 123 = '{' (a Channel JSON messsage).
+						// 123 = '{' (a Channel JSON message).
 						case 123:
-							this._processMessage(JSON.parse(nsPayload.toString('utf8')));
+							this._processMessage(JSON.parse(payload.toString('utf8')));
 							break;
 
 						// 68 = 'D' (a debug log).
 						case 68:
-							logger.debug(`[pid:${pid}] ${nsPayload.toString('utf8', 1)}`);
+							logger.debug(`[pid:${pid}] ${payload.toString('utf8', 1)}`);
 							break;
 
 						// 87 = 'W' (a warn log).
 						case 87:
-							logger.warn(`[pid:${pid}] ${nsPayload.toString('utf8', 1)}`);
+							logger.warn(`[pid:${pid}] ${payload.toString('utf8', 1)}`);
 							break;
 
 						// 69 = 'E' (an error log).
 						case 69:
-							logger.error(`[pid:${pid} ${nsPayload.toString('utf8', 1)}`);
+							logger.error(`[pid:${pid} ${payload.toString('utf8', 1)}`);
 							break;
 
 						// 88 = 'X' (a dump log).
 						case 88:
 							// eslint-disable-next-line no-console
-							console.log(nsPayload.toString('utf8', 1));
+							console.log(payload.toString('utf8', 1));
 							break;
 
 						default:
 							// eslint-disable-next-line no-console
 							console.warn(
 								`worker[pid:${pid}] unexpected data: %s`,
-								nsPayload.toString('utf8', 1));
+								payload.toString('utf8', 1));
 					}
 				}
 				catch (error)
@@ -155,17 +158,11 @@ export class Channel extends EnhancedEventEmitter
 						'received invalid message from the worker process: %s',
 						String(error));
 				}
+			}
 
-				// Remove the read payload from the buffer.
-				this._recvBuffer =
-					this._recvBuffer!.slice(netstring.nsLength(this._recvBuffer));
-
-				if (!this._recvBuffer.length)
-				{
-					this._recvBuffer = undefined;
-
-					return;
-				}
+			if (msgStart != 0)
+			{
+				this._recvBuffer = this._recvBuffer.slice(msgStart);
 			}
 		});
 
@@ -239,13 +236,15 @@ export class Channel extends EnhancedEventEmitter
 			throw new InvalidStateError('Channel closed');
 
 		const request = { id, method, internal, data };
-		const ns = netstring.nsWrite(JSON.stringify(request));
+		const payload = JSON.stringify(request);
 
-		if (Buffer.byteLength(ns) > NS_MESSAGE_MAX_LEN)
+		if (Buffer.byteLength(payload) > MESSAGE_MAX_LEN)
 			throw new Error('Channel request too big');
 
 		// This may throw if closed or remote side ended.
-		this._producerSocket.write(ns);
+		this._producerSocket.write(
+			Buffer.from(Uint32Array.of(Buffer.byteLength(payload)).buffer));
+		this._producerSocket.write(payload);
 
 		return new Promise((pResolve, pReject) =>
 		{