+++ worker/include/RTC/RTCP/FeedbackPsFir.hpp
@@ -35,6 +35,7 @@ namespace RTC
 			};
 
 		public:
+			static const size_t HeaderSize = 8;
 			static const FeedbackPs::MessageType messageType{ FeedbackPs::MessageType::FIR };
 
 		public:
@@ -62,7 +63,7 @@ namespace RTC
 			size_t Serialize(uint8_t* buffer) override;
 			size_t GetSize() const override
 			{
-				return sizeof(Header);
+				return FeedbackPsFirItem::HeaderSize;
 			}
 
 		private: