+++ worker/include/RTC/RTCP/FeedbackPsTst.hpp
@@ -12,8 +12,7 @@
   +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
   |                              SSRC                             |
   +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |  Seq nr.      |
-                  |  Reserved                           | Index   |
+  |  Seq nr.      |  Reserved                           | Index   |
   +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
  */
 
@@ -28,12 +27,19 @@ namespace RTC
 			struct Header
 			{
 				uint32_t ssrc;
-				uint32_t sequenceNumber : 8;
-				uint32_t reserved : 19;
-				uint32_t index : 5;
+				uint8_t sequenceNumber;
+				uint16_t reserved1;
+#if defined(MS_LITTLE_ENDIAN)
+				uint8_t index : 5;
+				uint8_t reserved2 : 3;
+#elif defined(MS_BIG_ENDIAN)
+				uint8_t reserved2 : 3;
+				uint8_t index : 5;
+#endif
 			};
 
 		public:
+			static const size_t HeaderSize{ 8 };
 			static const FeedbackPs::MessageType messageType;
 
 		public:
@@ -52,11 +58,11 @@ namespace RTC
 			}
 			uint8_t GetSequenceNumber() const
 			{
-				return static_cast<uint8_t>(this->header->sequenceNumber);
+				return this->header->sequenceNumber;
 			}
 			uint8_t GetIndex() const
 			{
-				return static_cast<uint8_t>(this->header->index);
+				return this->header->index;
 			}
 
 			/* Virtual methods inherited from FeedbackItem. */
@@ -65,7 +71,7 @@ namespace RTC
 			size_t Serialize(uint8_t* buffer) override;
 			size_t GetSize() const override
 			{
-				return sizeof(Header);
+				return FeedbackPsTstItem::HeaderSize;
 			}
 
 		private: