+++ worker/src/Channel/ChannelSocket.cpp
@@ -7,26 +7,22 @@
 #include <cmath>   // std::ceil()
 #include <cstdio>  // sprintf()
 #include <cstring> // std::memcpy(), std::memmove()
-extern "C"
-{
-#include <netstring.h>
-}
 
 namespace Channel
 {
 	/* Static. */
 
-	// netstring length for a 4194304 bytes payload.
-	static constexpr size_t NsMessageMaxLen{ 4194313 };
-	static constexpr size_t NsPayloadMaxLen{ 4194304 };
+	// Binary length for a 4194304 bytes payload.
+	static constexpr size_t MessageMaxLen{ 4194308 };
+	static constexpr size_t PayloadMaxLen{ 4194304 };
 
 	/* Instance methods. */
 	ChannelSocket::ChannelSocket(int consumerFd, int producerFd)
-	  : consumerSocket(consumerFd, NsMessageMaxLen, this), producerSocket(producerFd, NsMessageMaxLen)
+	  : consumerSocket(consumerFd, MessageMaxLen, this), producerSocket(producerFd, MessageMaxLen)
 	{
 		MS_TRACE_STD();
 
-		this->writeBuffer = static_cast<uint8_t*>(std::malloc(NsMessageMaxLen));
+		this->writeBuffer = static_cast<uint8_t*>(std::malloc(MessageMaxLen));
 	}
 
 	ChannelSocket::~ChannelSocket()
@@ -68,26 +64,26 @@ namespace Channel
 
 		std::string message = jsonMessage.dump();
 
-		if (message.length() > NsPayloadMaxLen)
+		if (message.length() > PayloadMaxLen)
 		{
-			MS_ERROR_STD("mesage too big");
+			MS_ERROR_STD("message too big");
 
 			return;
 		}
 
-		SendImpl(message.c_str(), message.length());
+		SendImpl(message.c_str(), static_cast<uint32_t>(message.length()));
 	}
 
-	void ChannelSocket::SendLog(char* message, size_t messageLen)
+	void ChannelSocket::SendLog(char* message, uint32_t messageLen)
 	{
 		MS_TRACE_STD();
 
 		if (this->closed)
 			return;
 
-		if (messageLen > NsPayloadMaxLen)
+		if (messageLen > PayloadMaxLen)
 		{
-			MS_ERROR_STD("mesage too big");
+			MS_ERROR_STD("message too big");
 
 			return;
 		}
@@ -95,30 +91,20 @@ namespace Channel
 		SendImpl(message, messageLen);
 	}
 
-	inline void ChannelSocket::SendImpl(const void* nsPayload, size_t nsPayloadLen)
+	inline void ChannelSocket::SendImpl(const void* payload, uint32_t payloadLen)
 	{
 		MS_TRACE_STD();
 
-		size_t nsNumLen;
+		std::memcpy(this->writeBuffer, &payloadLen, sizeof(uint32_t));
 
-		if (nsPayloadLen == 0)
+		if (payloadLen != 0)
 		{
-			nsNumLen             = 1;
-			this->writeBuffer[0] = '0';
-			this->writeBuffer[1] = ':';
-			this->writeBuffer[2] = ',';
-		}
-		else
-		{
-			nsNumLen = static_cast<size_t>(std::ceil(std::log10(static_cast<double>(nsPayloadLen) + 1)));
-			std::sprintf(reinterpret_cast<char*>(this->writeBuffer), "%zu:", nsPayloadLen);
-			std::memcpy(this->writeBuffer + nsNumLen + 1, nsPayload, nsPayloadLen);
-			this->writeBuffer[nsNumLen + nsPayloadLen + 1] = ',';
+			std::memcpy(this->writeBuffer + sizeof(uint32_t), payload, payloadLen);
 		}
 
-		size_t nsLen = nsNumLen + nsPayloadLen + 2;
+		size_t len = sizeof(uint32_t) + payloadLen;
 
-		this->producerSocket.Write(this->writeBuffer, nsLen);
+		this->producerSocket.Write(this->writeBuffer, len);
 	}
 
 	void ChannelSocket::OnConsumerSocketMessage(ConsumerSocket* /*consumerSocket*/, char* msg, size_t msgLen)
@@ -174,124 +160,48 @@ namespace Channel
 	{
 		MS_TRACE_STD();
 
+		size_t msgStart{ 0 };
+
 		// Be ready to parse more than a single message in a single chunk.
 		while (true)
 		{
 			if (IsClosed())
 				return;
 
-			size_t readLen = this->bufferDataLen - this->msgStart;
-			char* msgStart = nullptr;
-			size_t msgLen;
-			int nsRet = netstring_read(
-			  reinterpret_cast<char*>(this->buffer + this->msgStart), readLen, &msgStart, &msgLen);
+			size_t readLen = this->bufferDataLen - msgStart;
 
-			if (nsRet != 0)
+			if (readLen < sizeof(uint32_t))
 			{
-				switch (nsRet)
-				{
-					case NETSTRING_ERROR_TOO_SHORT:
-					{
-						// Check if the buffer is full.
-						if (this->bufferDataLen == this->bufferSize)
-						{
-							// First case: the incomplete message does not begin at position 0 of
-							// the buffer, so move the incomplete message to the position 0.
-							if (this->msgStart != 0)
-							{
-								std::memmove(this->buffer, this->buffer + this->msgStart, readLen);
-								this->msgStart      = 0;
-								this->bufferDataLen = readLen;
-							}
-							// Second case: the incomplete message begins at position 0 of the buffer.
-							// The message is too big, so discard it.
-							else
-							{
-								MS_ERROR_STD(
-								  "no more space in the buffer for the unfinished message being parsed, "
-								  "discarding it");
-
-								this->msgStart      = 0;
-								this->bufferDataLen = 0;
-							}
-						}
-
-						// Otherwise the buffer is not full, just wait.
-						return;
-					}
-
-					case NETSTRING_ERROR_TOO_LONG:
-					{
-						MS_ERROR_STD("NETSTRING_ERROR_TOO_LONG");
-
-						break;
-					}
-
-					case NETSTRING_ERROR_NO_COLON:
-					{
-						MS_ERROR_STD("NETSTRING_ERROR_NO_COLON");
-
-						break;
-					}
-
-					case NETSTRING_ERROR_NO_COMMA:
-					{
-						MS_ERROR_STD("NETSTRING_ERROR_NO_COMMA");
-
-						break;
-					}
-
-					case NETSTRING_ERROR_LEADING_ZERO:
-					{
-						MS_ERROR_STD("NETSTRING_ERROR_LEADING_ZERO");
-
-						break;
-					}
-
-					case NETSTRING_ERROR_NO_LENGTH:
-					{
-						MS_ERROR_STD("NETSTRING_ERROR_NO_LENGTH");
-
-						break;
-					}
-				}
-
-				// Error, so reset and exit the parsing loop.
-				this->msgStart      = 0;
-				this->bufferDataLen = 0;
-
-				return;
+				// Incomplete data.
+				break;
 			}
 
-			// If here it means that msgStart points to the beginning of a message
-			// with msgLen bytes length, so recalculate readLen.
-			readLen =
-			  reinterpret_cast<const uint8_t*>(msgStart) - (this->buffer + this->msgStart) + msgLen + 1;
+			uint32_t msgLen;
+			// Read message length.
+			std::memcpy(&msgLen, this->buffer + msgStart, sizeof(uint32_t));
 
-			this->listener->OnConsumerSocketMessage(this, msgStart, msgLen);
-
-			// If there is no more space available in the buffer and that is because
-			// the latest parsed message filled it, then empty the full buffer.
-			if ((this->msgStart + readLen) == this->bufferSize)
+			if (readLen < sizeof(uint32_t) + static_cast<size_t>(msgLen))
 			{
-				this->msgStart      = 0;
-				this->bufferDataLen = 0;
-			}
-			// If there is still space in the buffer, set the beginning of the next
-			// parsing to the next position after the parsed message.
-			else
-			{
-				this->msgStart += readLen;
+				// Incomplete data.
+				break;
 			}
 
-			// If there is more data in the buffer after the parsed message
-			// then parse again. Otherwise break here and wait for more data.
-			if (this->bufferDataLen > this->msgStart)
-			{
-				continue;
-			}
+			this->listener->OnConsumerSocketMessage(
+			  this,
+			  reinterpret_cast<char*>(this->buffer + msgStart + sizeof(uint32_t)),
+			  static_cast<size_t>(msgLen));
 
-			break;
+			msgStart += sizeof(uint32_t) + static_cast<size_t>(msgLen);
+		}
+
+		if (msgStart != 0)
+		{
+			this->bufferDataLen = this->bufferDataLen - msgStart;
+
+			if (this->bufferDataLen != 0)
+			{
+				std::memmove(this->buffer, this->buffer + msgStart, this->bufferDataLen);
+			}
 		}
 	}