+++ worker/src/RTC/TransportCongestionControlClient.cpp
@@ -16,6 +16,7 @@ namespace RTC
 	static constexpr float MaxBitrateIncrementFactor{ 1.35f };
 	static constexpr float MaxPaddingBitrateFactor{ 0.85f };
 	static constexpr uint64_t AvailableBitrateEventInterval{ 2000u }; // In ms.
+	static constexpr size_t PacketLossHistogramLength{ 24 };
 
 	/* Instance methods. */
 
@@ -163,7 +164,54 @@ namespace RTC
 	{
 		MS_TRACE();
 
+		// Update packet loss history.
+		size_t expected_packets = feedback->GetPacketStatusCount();
+		size_t lost_packets     = 0;
+		for (const auto& result : feedback->GetPacketResults())
+		{
+			if (!result.received)
+				lost_packets += 1;
+		}
+		this->UpdatePacketLoss(static_cast<double>(lost_packets) / expected_packets);
+
 		this->rtpTransportControllerSend->OnTransportFeedback(*feedback);
+	}
+
+	void TransportCongestionControlClient::UpdatePacketLoss(double packetLoss)
+	{
+		// Add the score into the histogram.
+		if (this->packetLossHistory.size() == PacketLossHistogramLength)
+			this->packetLossHistory.pop_front();
+
+		this->packetLossHistory.push_back(packetLoss);
+
+		/*
+		 * Scoring mechanism is a weighted average.
+		 *
+		 * The more recent the score is, the more weight it has.
+		 * The oldest score has a weight of 1 and subsequent scores weight is
+		 * increased by one sequentially.
+		 *
+		 * Ie:
+		 * - scores: [1,2,3,4]
+		 * - this->scores = ((1) + (2+2) + (3+3+3) + (4+4+4+4)) / 10 = 2.8 => 3
+		 */
+
+		size_t weight{ 0 };
+		size_t samples{ 0 };
+		double totalPacketLoss{ 0 };
+
+		for (auto packetLossEntry : this->packetLossHistory)
+		{
+			weight++;
+			samples += weight;
+			totalPacketLoss += weight * packetLossEntry;
+		}
+
+		// clang-tidy "thinks" that this can lead to division by zero but we are
+		// smarter.
+		// NOLINTNEXTLINE(clang-analyzer-core.DivideZero)
+		this->packetLoss = totalPacketLoss / samples;
 	}
 
 	void TransportCongestionControlClient::SetMaxOutgoingBitrate(uint32_t maxBitrate)
@@ -243,6 +291,13 @@ namespace RTC
 		MS_TRACE();
 
 		return this->bitrates.availableBitrate;
+	}
+
+	double TransportCongestionControlClient::GetPacketLoss() const
+	{
+		MS_TRACE();
+
+		return this->packetLoss;
 	}
 
 	void TransportCongestionControlClient::RescheduleNextAvailableBitrateEvent()