+++ rust/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+# Unreleased
+
+* Fix for receiving data over payload channel
+
+# 0.8.5
+
+* Fix types for `round_trip_time` and `bitrate_by_layer` fields `ProducerStat` and `ConsumerStat`
+* Accumulation of worker fixes
+
 # 0.8.4
 
 * Add Active Speaker Observer to prelude